refactor(server): Move Nozl related files to a separate directory
