+++ nozl/eventstream/eventstream.go
@@ -8,7 +8,7 @@ import (
 	"github.com/avast/retry-go/v4"
 	"github.com/nats-io/nats.go"
 
-	"github.com/nats-io/nats-server/v2/server/backend/shared"
+	"github.com/nats-io/nats-server/v2/nozl/shared"
 )
 
 type (
