+++ nozl/eventstream/handler.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/labstack/echo/v4"
-	"github.com/nats-io/nats-server/v2/server/backend/shared"
+	"github.com/nats-io/nats-server/v2/nozl/shared"
 )
 
 func SendMessageHandler(ctx echo.Context) error {
