+++ nozl/eventstream/message.go
