+++ nozl/eventstream/message_test.go
