+++ nozl/shared/constants.go
