+++ nozl/shared/stack.go
